+++ test/services/blob/blobservice-tests.js
@@ -1155,6 +1155,20 @@ describe('BlobService', function () {
         blobUrl = blobServiceassert.getUrl(containerName, blobName);
         assert.equal(blobUrl, 'https://host.com:88/' + containerName + '/' + blobName);
 
+        blobServiceassert = azure.createBlobService('storageAccount', 'storageAccessKey');
+        blobServiceassert.setHost({primaryHost: 'host.com:88/account'});
+        blobUrl = blobServiceassert.getUrl(containerName);
+        assert.equal(blobUrl, 'https://host.com:88/account/' + containerName);
+        blobUrl = blobServiceassert.getUrl(containerName, blobName);
+        assert.equal(blobUrl, 'https://host.com:88/account/' + containerName + '/' + blobName);
+
+        blobServiceassert = azure.createBlobService('storageAccount', 'storageAccessKey');
+        blobServiceassert.setHost({primaryHost: 'host.com:88/account'});
+        blobUrl = blobServiceassert.getUrl(containerName);
+        assert.equal(blobUrl, 'https://host.com:88/account/' + containerName);
+        blobUrl = blobServiceassert.getUrl(containerName, blobName, null, true, '2016-10-11T11:03:40Z');
+        assert.equal(blobUrl, 'https://host.com:88/account/' + containerName + '/' + blobName + '?snapshot=2016-10-11T11%3A03%3A40Z');
+
         done();
       });
 
@@ -1162,7 +1176,7 @@ describe('BlobService', function () {
         var containerName = 'container';
         var blobName = 'blob';
 
-        var blobServiceassert = azure.createBlobService('storageAccount', 'storageAccessKey', 'host.com:80');
+        var blobServiceassert = azure.createBlobService('storageAccount', 'storageAccessKey', {primaryHost: 'https://host.com:80/', secondaryHost: 'https://host-secondary.com:80/'});
 
         var sharedAccessPolicy = {
           AccessPolicy: {
@@ -1171,7 +1185,8 @@ describe('BlobService', function () {
           }
         };
 
-        var blobUrl = blobServiceassert.getUrl(containerName, blobName, blobServiceassert.generateSharedAccessSignature(containerName, blobName, sharedAccessPolicy));
+        var sasToken = blobServiceassert.generateSharedAccessSignature(containerName, blobName, sharedAccessPolicy);
+        var blobUrl = blobServiceassert.getUrl(containerName, blobName, sasToken);
 
         var parsedUrl = url.parse(blobUrl);
         assert.strictEqual(parsedUrl.protocol, 'https:');
@@ -1179,6 +1194,15 @@ describe('BlobService', function () {
         assert.strictEqual(parsedUrl.hostname, 'host.com');
         assert.strictEqual(parsedUrl.pathname, '/' + containerName + '/' + blobName);
         assert.strictEqual(parsedUrl.query, 'se=2011-10-12T11%3A53%3A40Z&spr=https&sv=2015-04-05&sr=b&sig=wrbAp%2BxDtX5rwkAk8IkxspEHid2DiwE3JqVr%2BNDA2Bk%3D');
+
+        blobUrl = blobServiceassert.getUrl(containerName, blobName, sasToken, false, '2016-10-11T11:03:40Z');
+
+        var parsedUrl = url.parse(blobUrl);
+        assert.strictEqual(parsedUrl.protocol, 'https:');
+        assert.strictEqual(parsedUrl.port, '80');
+        assert.strictEqual(parsedUrl.hostname, 'host-secondary.com');
+        assert.strictEqual(parsedUrl.pathname, '/' + containerName + '/' + blobName);
+        assert.strictEqual(parsedUrl.query, 'se=2011-10-12T11%3A53%3A40Z&spr=https&sv=2015-04-05&sr=b&sig=wrbAp%2BxDtX5rwkAk8IkxspEHid2DiwE3JqVr%2BNDA2Bk%3D&snapshot=2016-10-11T11%3A03%3A40Z');
 
         done();
       });